Extract workflow engine helpers into shared modules
- Move context-step prompt and hint logic into dedicated helpers
- Split assess-signal shaping out of review loop executor
- Consolidate workflow test setup into shared helpers
