+++ packages/backend/test/workflowEngine/context-steps.test.ts
@@ -1,84 +1,17 @@
-// @ts-nocheck
-/* eslint-disable @typescript-eslint/no-unused-vars */
 /**
- * @description: Verifies workflow-engine transition and limit invariants used by backend orchestration.
+ * @description: Verifies context-step execution, merge ordering, and fail-open behavior.
  * @footnote-scope: test
- * @footnote-module: WorkflowEngineTests
- * @footnote-risk: medium - Missing coverage can allow transition or budget regressions in shared orchestration logic.
- * @footnote-ethics: high - Workflow bounds and legality checks enforce auditable, fail-open-safe model control.
+ * @footnote-module: WorkflowEngineContextStepTests
+ * @footnote-risk: medium - Context-step regressions can misapply tools or block generation.
+ * @footnote-ethics: high - Context provenance and clarification behavior must stay auditable.
  */
 import test from 'node:test';
 import assert from 'node:assert/strict';
-import { readFileSync } from 'node:fs';
-import { dirname, join } from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-import {
-    applyStepExecutionToState,
-    buildPlannerStepRecord,
-    createInitialWorkflowState,
-    isWorkflowTransitionAllowed,
-    checkExecutionLimits,
-    mapLimitExhaustionToTerminationReason,
-    runBoundedReviewWorkflow,
-    type ExecutionLimits,
-    type WorkflowRunPolicy,
-} from '../../src/services/workflowEngine.js';
 import type {
     GenerationRuntime,
     RuntimeMessage,
 } from '@footnote/agent-runtime';
-import type { ConversationContextEnvelope } from '../../src/services/conversationContextService.js';
-
-const permissivePolicy: WorkflowRunPolicy = {
-    enablePlanning: true,
-    enableToolUse: true,
-    enableReplanning: true,
-    enableAssessment: true,
-    enableRevision: true,
-};
-
-const strictPolicy: WorkflowRunPolicy = {
-    enablePlanning: false,
-    enableToolUse: false,
-    enableReplanning: false,
-    enableAssessment: false,
-    enableRevision: false,
-};
-
-const createLimits = (): ExecutionLimits => ({
-    maxWorkflowSteps: 5,
-    maxToolCalls: 2,
-    maxDeliberationCalls: 3,
-    maxTokensTotal: 100,
-    maxDurationMs: 1000,
-});
-
-const TEST_CONTEXT_ENVELOPE: ConversationContextEnvelope = {
-    participants: [],
-    turns: [],
-    diagnostics: {
-        surface: 'web',
-        totalInputMessages: 0,
-        projectedMessageCount: 0,
-        trimmedMessageCount: 0,
-        sanitizedTimestampCount: 0,
-        projectedSpeakerLabelCount: 0,
-    },
-};
-
-const runBoundedReviewWorkflowForTest = (
-    input: Omit<
-        Parameters<typeof runBoundedReviewWorkflow>[0],
-        'contextEnvelope'
-    > & {
-        contextEnvelope?: ConversationContextEnvelope;
-    }
-): ReturnType<typeof runBoundedReviewWorkflow> =>
-    runBoundedReviewWorkflow({
-        ...input,
-        contextEnvelope: input.contextEnvelope ?? TEST_CONTEXT_ENVELOPE,
-    });
+import { runBoundedReviewWorkflowForTest } from './helpers.js';
 
 test('runBoundedReviewWorkflow does not emit concrete tool steps in current engine-owned review path', async () => {
     const generationRuntime: GenerationRuntime = {
