+++ packages/backend/test/workflowEngine/limits.test.ts
@@ -1,50 +1,17 @@
-// @ts-nocheck
-/* eslint-disable @typescript-eslint/no-unused-vars */
 /**
- * @description: Verifies workflow-engine transition and limit invariants used by backend orchestration.
+ * @description: Verifies workflow execution-limit checks and termination mappings.
  * @footnote-scope: test
- * @footnote-module: WorkflowEngineTests
- * @footnote-risk: medium - Missing coverage can allow transition or budget regressions in shared orchestration logic.
- * @footnote-ethics: high - Workflow bounds and legality checks enforce auditable, fail-open-safe model control.
+ * @footnote-module: WorkflowEngineLimitTests
+ * @footnote-risk: medium - Limit regressions can cause runaway or premature stops.
+ * @footnote-ethics: high - Bound enforcement is core to safe execution control.
  */
 import test from 'node:test';
 import assert from 'node:assert/strict';
-import { readFileSync } from 'node:fs';
-import { dirname, join } from 'node:path';
-import { fileURLToPath } from 'node:url';
-
 import {
-    applyStepExecutionToState,
-    buildPlannerStepRecord,
-    createInitialWorkflowState,
-    isWorkflowTransitionAllowed,
     checkExecutionLimits,
     mapLimitExhaustionToTerminationReason,
-    runBoundedReviewWorkflow,
     type ExecutionLimits,
-    type WorkflowRunPolicy,
 } from '../../src/services/workflowEngine.js';
-import type {
-    GenerationRuntime,
-    RuntimeMessage,
-} from '@footnote/agent-runtime';
-import type { ConversationContextEnvelope } from '../../src/services/conversationContextService.js';
-
-const permissivePolicy: WorkflowRunPolicy = {
-    enablePlanning: true,
-    enableToolUse: true,
-    enableReplanning: true,
-    enableAssessment: true,
-    enableRevision: true,
-};
-
-const strictPolicy: WorkflowRunPolicy = {
-    enablePlanning: false,
-    enableToolUse: false,
-    enableReplanning: false,
-    enableAssessment: false,
-    enableRevision: false,
-};
 
 const createLimits = (): ExecutionLimits => ({
     maxWorkflowSteps: 5,
@@ -53,32 +20,6 @@ const createLimits = (): ExecutionLimits => ({
     maxTokensTotal: 100,
     maxDurationMs: 1000,
 });
-
-const TEST_CONTEXT_ENVELOPE: ConversationContextEnvelope = {
-    participants: [],
-    turns: [],
-    diagnostics: {
-        surface: 'web',
-        totalInputMessages: 0,
-        projectedMessageCount: 0,
-        trimmedMessageCount: 0,
-        sanitizedTimestampCount: 0,
-        projectedSpeakerLabelCount: 0,
-    },
-};
-
-const runBoundedReviewWorkflowForTest = (
-    input: Omit<
-        Parameters<typeof runBoundedReviewWorkflow>[0],
-        'contextEnvelope'
-    > & {
-        contextEnvelope?: ConversationContextEnvelope;
-    }
-): ReturnType<typeof runBoundedReviewWorkflow> =>
-    runBoundedReviewWorkflow({
-        ...input,
-        contextEnvelope: input.contextEnvelope ?? TEST_CONTEXT_ENVELOPE,
-    });
 
 test('checkExecutionLimits reports each exhausted limit key', () => {
     const limits = createLimits();
